+++ test/suite/systems/test_hamiltonian_getter.jl
@@ -3,6 +3,7 @@ module TestHamiltonianGetter
 import Test
 import ADTypes
 import DifferentiationInterface
+import ForwardDiff  # triggers DifferentiationInterfaceForwardDiff (provides PushforwardFast)
 import CTFlows.Common
 import CTBase.Traits
 import CTBase.Data